test(api-core): move expected power shelf tests to owning suites
