+++ crates/api-core/src/tests/common/api_fixtures/site_explorer.rs
@@ -2190,38 +2190,3 @@ pub async fn create_expected_switches(
     }
     created
 }
-
-/// create_expected_power_shelves seeds 6 expected power shelves into the
-/// database, replacing the create_expected_power_shelf.sql fixture.
-pub async fn create_expected_power_shelves(
-    txn: &mut sqlx::PgConnection,
-) -> Vec<model::expected_power_shelf::ExpectedPowerShelf> {
-    use model::expected_power_shelf::ExpectedPowerShelf;
-    use model::metadata::Metadata;
-
-    use crate::test_support::mac_address_pool::EXPECTED_POWER_SHELF_BMC_MAC_ADDRESS_POOL;
-
-    let mut created = Vec::new();
-    for i in 0..6 {
-        let power_shelf = ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: EXPECTED_POWER_SHELF_BMC_MAC_ADDRESS_POOL.allocate(),
-            serial_number: format!("PS-SN-{:03}", i + 1),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "Pwd2023x0x0x0x0x7".into(),
-            bmc_ip_address: if (3..=4).contains(&i) {
-                Some(format!("192.168.1.{}", 100 + i - 3).parse().unwrap())
-            } else {
-                None
-            },
-            metadata: Metadata::default(),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        };
-        let result = db::expected_power_shelf::create(txn, power_shelf)
-            .await
-            .expect("unable to create expected power shelf");
-        created.push(result);
-    }
-    created
-}
