+++ crates/api-core/src/tests/mod.rs
@@ -30,7 +30,6 @@ mod dpu_remediation;
 mod dpu_reprovisioning;
 mod dynamic_config;
 mod expected_machine;
-mod expected_power_shelf;
 mod expected_switch;
 mod explored_endpoint_find;
 mod extension_service;
