+++ crates/api-core/tests/integration/expected_power_shelf_crud.rs
@@ -14,19 +14,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 use std::default::Default;
 
+use carbide_test_harness::prelude::*;
 use carbide_uuid::rack::RackId;
-use common::api_fixtures::create_test_env;
-use common::api_fixtures::site_explorer::create_expected_power_shelves;
 use mac_address::MacAddress;
 use rpc::forge::forge_server::Forge;
 use rpc::forge::{ExpectedPowerShelfList, ExpectedPowerShelfRequest};
 use uuid::Uuid;
 
-use crate::tests::common;
+use crate::expected_power_shelf::{create_expected_power_shelves, create_test_env};
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_add_expected_power_shelf(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
 
@@ -78,7 +78,7 @@ async fn test_add_expected_power_shelf(pool: sqlx::PgPool) {
             bmc_retain_credentials: None,
         },
     ] {
-        env.api
+        env.api()
             .add_expected_power_shelf(tonic::Request::new(expected_power_shelf.clone()))
             .await
             .expect("unable to add expected power shelf ");
@@ -89,7 +89,7 @@ async fn test_add_expected_power_shelf(pool: sqlx::PgPool) {
         };
 
         let mut retrieved_expected_power_shelf = env
-            .api
+            .api()
             .get_expected_power_shelf(tonic::Request::new(expected_power_shelf_query))
             .await
             .expect("unable to retrieve expected power shelf ")
@@ -112,7 +112,7 @@ async fn test_add_expected_power_shelf(pool: sqlx::PgPool) {
     }
 }
 
-#[crate::sqlx_test]
+#[sqlx_test]
 async fn test_delete_expected_power_shelf(pool: sqlx::PgPool) {
     let mut conn = pool.acquire().await.unwrap();
     let shelves = create_expected_power_shelves(&mut conn).await;
@@ -120,7 +120,7 @@ async fn test_delete_expected_power_shelf(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
 
     let expected_power_shelf_count = env
-        .api
+        .api()
         .get_all_expected_power_shelves(tonic::Request::new(()))
         .await
         .expect("unable to get all expected power shelves")
@@ -132,14 +132,14 @@ async fn test_delete_expected_power_shelf(pool: sqlx::PgPool) {
         bmc_mac_address: shelves[1].bmc_mac_address.to_string(),
         expected_power_shelf_id: None,
     };
-    env.api
+    env.api()
         .delete_expected_power_shelf(tonic::Request::new(expected_power_shelf_query))
         .await
         .expect("unable to delete expected power shelf ")
         .into_inner();
 
     let new_expected_power_shelf_count = env
-        .api
+        .api()
         .get_all_expected_power_shelves(tonic::Request::new(()))
         .await
         .expect("unable to get all expected power shelves")
@@ -153,7 +153,7 @@ async fn test_delete_expected_power_shelf(pool: sqlx::PgPool) {
     );
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_delete_expected_power_shelf_error(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
     let bmc_mac_address: MacAddress = "2A:2B:2C:2D:2E:2F".parse().unwrap();
@@ -163,7 +163,7 @@ async fn test_delete_expected_power_shelf_error(pool: sqlx::PgPool) {
     };
 
     let err = env
-        .api
+        .api()
         .delete_expected_power_shelf(tonic::Request::new(expected_power_shelf_request))
         .await
         .unwrap_err();
@@ -174,7 +174,7 @@ async fn test_delete_expected_power_shelf_error(pool: sqlx::PgPool) {
     );
 }
 
-#[crate::sqlx_test]
+#[sqlx_test]
 async fn test_update_expected_power_shelf(pool: sqlx::PgPool) {
     let mut conn = pool.acquire().await.unwrap();
     let shelves = create_expected_power_shelves(&mut conn).await;
@@ -230,14 +230,14 @@ async fn test_update_expected_power_shelf(pool: sqlx::PgPool) {
             bmc_retain_credentials: None,
         },
     ] {
-        env.api
+        env.api()
             .update_expected_power_shelf(tonic::Request::new(updated_power_shelf.clone()))
             .await
             .expect("unable to update expected power shelf ")
             .into_inner();
 
         let mut retrieved_expected_power_shelf = env
-            .api
+            .api()
             .get_expected_power_shelf(tonic::Request::new(ExpectedPowerShelfRequest {
                 bmc_mac_address: bmc_mac_address.to_string(),
                 expected_power_shelf_id: None,
@@ -263,7 +263,7 @@ async fn test_update_expected_power_shelf(pool: sqlx::PgPool) {
     }
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_update_expected_power_shelf_error(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
     let bmc_mac_address: MacAddress = "2A:2B:2C:2D:2E:2F".parse().unwrap();
@@ -280,7 +280,7 @@ async fn test_update_expected_power_shelf_error(pool: sqlx::PgPool) {
     };
 
     let err = env
-        .api
+        .api()
         .update_expected_power_shelf(tonic::Request::new(expected_power_shelf.clone()))
         .await
         .unwrap_err();
@@ -292,14 +292,14 @@ async fn test_update_expected_power_shelf_error(pool: sqlx::PgPool) {
     );
 }
 
-#[crate::sqlx_test]
+#[sqlx_test]
 async fn test_delete_all_expected_power_shelves(pool: sqlx::PgPool) {
     let mut conn = pool.acquire().await.unwrap();
     create_expected_power_shelves(&mut conn).await;
     drop(conn);
     let env = create_test_env(pool).await;
     let mut expected_power_shelf_count = env
-        .api
+        .api()
         .get_all_expected_power_shelves(tonic::Request::new(()))
         .await
         .expect("unable to get all expected power shelves")
@@ -309,14 +309,14 @@ async fn test_delete_all_expected_power_shelves(pool: sqlx::PgPool) {
 
     assert_eq!(expected_power_shelf_count, 6);
 
-    env.api
+    env.api()
         .delete_all_expected_power_shelves(tonic::Request::new(()))
         .await
         .expect("unable to delete all expected power shelves")
         .into_inner();
 
     expected_power_shelf_count = env
-        .api
+        .api()
         .get_all_expected_power_shelves(tonic::Request::new(()))
         .await
         .expect("unable to get all expected power shelves")
@@ -327,14 +327,14 @@ async fn test_delete_all_expected_power_shelves(pool: sqlx::PgPool) {
     assert_eq!(expected_power_shelf_count, 0);
 }
 
-#[crate::sqlx_test]
+#[sqlx_test]
 async fn test_replace_all_expected_power_shelves(pool: sqlx::PgPool) {
     let mut conn = pool.acquire().await.unwrap();
     create_expected_power_shelves(&mut conn).await;
     drop(conn);
     let env = create_test_env(pool).await;
     let expected_power_shelf_count = env
-        .api
+        .api()
         .get_all_expected_power_shelves(tonic::Request::new(()))
         .await
         .expect("unable to get all expected power shelves")
@@ -379,14 +379,14 @@ async fn test_replace_all_expected_power_shelves(pool: sqlx::PgPool) {
         .expected_power_shelves
         .push(expected_power_shelf_2.clone());
 
-    env.api
+    env.api()
         .replace_all_expected_power_shelves(tonic::Request::new(expected_power_shelf_list))
         .await
         .expect("unable to replace all expected power shelves")
         .into_inner();
 
     let expected_power_shelves = env
-        .api
+        .api()
         .get_all_expected_power_shelves(tonic::Request::new(()))
         .await
         .expect("unable to get all expected power shelves")
@@ -407,7 +407,7 @@ async fn test_replace_all_expected_power_shelves(pool: sqlx::PgPool) {
     );
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_get_expected_power_shelf_error(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
     let bmc_mac_address: MacAddress = "2A:2B:2C:2D:2E:2F".parse().unwrap();
@@ -417,7 +417,7 @@ async fn test_get_expected_power_shelf_error(pool: sqlx::PgPool) {
     };
 
     let err = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(expected_power_shelf_query))
         .await
         .unwrap_err();
@@ -429,14 +429,14 @@ async fn test_get_expected_power_shelf_error(pool: sqlx::PgPool) {
     );
 }
 
-#[crate::sqlx_test]
+#[sqlx_test]
 async fn test_get_linked_expected_power_shelves_unseen(pool: sqlx::PgPool) {
     let mut conn = pool.acquire().await.unwrap();
     create_expected_power_shelves(&mut conn).await;
     drop(conn);
     let env = create_test_env(pool).await;
     let out = env
-        .api
+        .api()
         .get_all_expected_power_shelves_linked(tonic::Request::new(()))
         .await
         .unwrap()
@@ -451,7 +451,7 @@ async fn test_get_linked_expected_power_shelves_unseen(pool: sqlx::PgPool) {
     );
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_add_expected_power_shelf_with_ip(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
     let bmc_mac_address: MacAddress = "3A:3B:3C:3D:3E:3F".parse().unwrap();
@@ -467,7 +467,7 @@ async fn test_add_expected_power_shelf_with_ip(pool: sqlx::PgPool) {
         bmc_retain_credentials: None,
     };
 
-    env.api
+    env.api()
         .add_expected_power_shelf(tonic::Request::new(expected_power_shelf.clone()))
         .await
         .expect("unable to add expected power shelf ");
@@ -478,7 +478,7 @@ async fn test_add_expected_power_shelf_with_ip(pool: sqlx::PgPool) {
     };
 
     let retrieved_expected_power_shelf = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(expected_power_shelf_query))
         .await
         .expect("unable to retrieve expected power shelf ")
@@ -492,35 +492,17 @@ async fn test_add_expected_power_shelf_with_ip(pool: sqlx::PgPool) {
     assert_eq!(retrieved_expected_power_shelf.bmc_ip_address, "10.0.0.100");
 }
 
-#[crate::sqlx_test]
-async fn test_with_ip_addresses(pool: sqlx::PgPool) -> Result<(), Box<dyn std::error::Error>> {
-    let mut conn = pool.acquire().await.unwrap();
-    let shelves = create_expected_power_shelves(&mut conn).await;
-    drop(conn);
-
-    // Shelves at indices 3 and 4 are created with IP addresses
-    assert_eq!(
-        shelves[3].bmc_ip_address,
-        Some("192.168.1.100".parse().unwrap())
-    );
-    assert_eq!(
-        shelves[4].bmc_ip_address,
-        Some("192.168.1.101".parse().unwrap())
-    );
-
-    Ok(())
-}
-
-#[crate::sqlx_test]
+#[sqlx_test]
 async fn test_update_expected_power_shelf_ip_address(pool: sqlx::PgPool) {
     let mut conn = pool.acquire().await.unwrap();
     let shelves = create_expected_power_shelves(&mut conn).await;
+
     drop(conn);
     let env = create_test_env(pool).await;
 
     let shelf_mac = shelves[1].bmc_mac_address.to_string();
     let mut eps1 = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelfRequest {
             bmc_mac_address: shelf_mac.clone(),
             expected_power_shelf_id: None,
@@ -531,14 +513,14 @@ async fn test_update_expected_power_shelf_ip_address(pool: sqlx::PgPool) {
 
     eps1.bmc_ip_address = "172.16.0.50".to_string();
 
-    env.api
+    env.api()
         .update_expected_power_shelf(tonic::Request::new(eps1.clone()))
         .await
         .expect("unable to update")
         .into_inner();
 
     let eps2 = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelfRequest {
             bmc_mac_address: shelf_mac,
             expected_power_shelf_id: None,
@@ -551,7 +533,7 @@ async fn test_update_expected_power_shelf_ip_address(pool: sqlx::PgPool) {
     assert_eq!(eps2.bmc_ip_address, "172.16.0.50");
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_get_expected_power_shelf_by_id(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
 
@@ -570,7 +552,7 @@ async fn test_get_expected_power_shelf_by_id(pool: sqlx::PgPool) {
         bmc_retain_credentials: None,
     };
 
-    env.api
+    env.api()
         .add_expected_power_shelf(tonic::Request::new(expected_power_shelf.clone()))
         .await
         .expect("unable to add expected power shelf");
@@ -583,7 +565,7 @@ async fn test_get_expected_power_shelf_by_id(pool: sqlx::PgPool) {
         }),
     };
     let retrieved = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(get_req))
         .await
         .expect("unable to retrieve by id")
@@ -598,7 +580,7 @@ async fn test_get_expected_power_shelf_by_id(pool: sqlx::PgPool) {
     assert_eq!(retrieved.bmc_ip_address, "10.0.0.50");
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_delete_expected_power_shelf_by_id(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
 
@@ -617,7 +599,7 @@ async fn test_delete_expected_power_shelf_by_id(pool: sqlx::PgPool) {
         bmc_retain_credentials: None,
     };
 
-    env.api
+    env.api()
         .add_expected_power_shelf(tonic::Request::new(expected_power_shelf.clone()))
         .await
         .expect("unable to add expected power shelf");
@@ -629,7 +611,7 @@ async fn test_delete_expected_power_shelf_by_id(pool: sqlx::PgPool) {
             value: provided_id.clone(),
         }),
     };
-    env.api
+    env.api()
         .delete_expected_power_shelf(tonic::Request::new(del_req))
         .await
         .expect("unable to delete by id");
@@ -642,7 +624,7 @@ async fn test_delete_expected_power_shelf_by_id(pool: sqlx::PgPool) {
         }),
     };
     let err = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(get_req))
         .await
         .unwrap_err();
@@ -653,7 +635,7 @@ async fn test_delete_expected_power_shelf_by_id(pool: sqlx::PgPool) {
     );
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_update_expected_power_shelf_by_id(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
 
@@ -672,7 +654,7 @@ async fn test_update_expected_power_shelf_by_id(pool: sqlx::PgPool) {
         bmc_retain_credentials: None,
     };
 
-    env.api
+    env.api()
         .add_expected_power_shelf(tonic::Request::new(expected_power_shelf.clone()))
         .await
         .expect("unable to add expected power shelf");
@@ -681,7 +663,7 @@ async fn test_update_expected_power_shelf_by_id(pool: sqlx::PgPool) {
     expected_power_shelf.bmc_username = "ADMIN_UPDATED".into();
     expected_power_shelf.shelf_serial_number = "PS-UPD-ID-002".into();
     expected_power_shelf.bmc_ip_address = "172.16.0.99".into();
-    env.api
+    env.api()
         .update_expected_power_shelf(tonic::Request::new(expected_power_shelf.clone()))
         .await
         .expect("unable to update by id");
@@ -694,7 +676,7 @@ async fn test_update_expected_power_shelf_by_id(pool: sqlx::PgPool) {
         }),
     };
     let retrieved = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(get_req))
         .await
         .expect("unable to get after update by id")
@@ -709,7 +691,7 @@ async fn test_update_expected_power_shelf_by_id(pool: sqlx::PgPool) {
     assert_eq!(retrieved.bmc_ip_address, "172.16.0.99");
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_create_expected_power_shelf_with_explicit_id(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
 
@@ -728,7 +710,7 @@ async fn test_create_expected_power_shelf_with_explicit_id(pool: sqlx::PgPool) {
         bmc_retain_credentials: None,
     };
 
-    env.api
+    env.api()
         .add_expected_power_shelf(tonic::Request::new(expected_power_shelf.clone()))
         .await
         .expect("unable to add expected power shelf with explicit id");
@@ -739,7 +721,7 @@ async fn test_create_expected_power_shelf_with_explicit_id(pool: sqlx::PgPool) {
         expected_power_shelf_id: None,
     };
     let retrieved = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(get_req))
         .await
         .expect("unable to retrieve expected power shelf")
@@ -751,7 +733,7 @@ async fn test_create_expected_power_shelf_with_explicit_id(pool: sqlx::PgPool) {
     );
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_create_expected_power_shelf_auto_generates_id(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
 
@@ -767,7 +749,7 @@ async fn test_create_expected_power_shelf_auto_generates_id(pool: sqlx::PgPool) 
         bmc_retain_credentials: None,
     };
 
-    env.api
+    env.api()
         .add_expected_power_shelf(tonic::Request::new(expected_power_shelf.clone()))
         .await
         .expect("unable to add expected power shelf without id");
@@ -778,7 +760,7 @@ async fn test_create_expected_power_shelf_auto_generates_id(pool: sqlx::PgPool) 
         expected_power_shelf_id: None,
     };
     let retrieved = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(get_req))
         .await
         .expect("unable to retrieve expected power shelf")
@@ -799,7 +781,7 @@ async fn test_create_expected_power_shelf_auto_generates_id(pool: sqlx::PgPool) 
     );
 }
 
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_get_expected_power_shelf_by_id_not_found(pool: sqlx::PgPool) {
     let env = create_test_env(pool).await;
 
@@ -812,7 +794,7 @@ async fn test_get_expected_power_shelf_by_id_not_found(pool: sqlx::PgPool) {
     };
 
     let err = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(get_req))
         .await
         .unwrap_err();
@@ -823,467 +805,16 @@ async fn test_get_expected_power_shelf_by_id_not_found(pool: sqlx::PgPool) {
     );
 }
 
-/// When an expected power shelf is created with a bmc_ip_address, test to make
-/// sure a machine_interface is pre-allocated with a static address in the DB.
-/// Site explorer then just picks it up naturally from the underlay interface query.
-#[crate::sqlx_test()]
-async fn test_add_with_bmc_ip_creates_static_interface(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let bmc_mac: MacAddress = "4A:4B:4C:4D:4E:4F".parse().unwrap();
-    let bmc_ip = "192.0.2.180";
-
-    env.api
-        .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-STATIC-001".into(),
-            bmc_ip_address: bmc_ip.into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // Add doesn't preallocate inline; mimic what site-explorer does on the next iteration --
-    // materialize the static BMC interface for this entity.
-    carbide_site_explorer::try_preallocate_one(
-        &env.pool,
-        bmc_mac,
-        bmc_ip.parse().unwrap(),
-        model::machine_interface::InterfaceType::Bmc,
-        "expected_power_shelf BMC",
-        None,
-    )
-    .await;
-
-    let mut txn = env.pool.begin().await?;
-    let interfaces = db::machine_interface::find_by_mac_address(&mut *txn, bmc_mac).await?;
-    assert_eq!(
-        interfaces.len(),
-        1,
-        "should have one interface for the BMC MAC"
-    );
-
-    let iface = &interfaces[0];
-    assert!(
-        iface.addresses.contains(&bmc_ip.parse().unwrap()),
-        "interface should have the static BMC IP"
-    );
-
-    // Verify the address is a static allocation type.
-    let addrs = db::machine_interface_address::find_for_interface(&mut txn, iface.id).await?;
-    assert_eq!(addrs.len(), 1);
-    assert_eq!(
-        addrs[0].address,
-        bmc_ip.parse::<std::net::IpAddr>().unwrap()
-    );
-    assert_eq!(
-        addrs[0].allocation_type,
-        model::allocation_type::AllocationType::Static
-    );
-
-    Ok(())
-}
-
-/// When an expected power shelf is created WITHOUT a bmc_ip_address,
-/// no machine_interface should be created.
-#[crate::sqlx_test()]
-async fn test_add_without_bmc_ip_creates_no_interface(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let bmc_mac: MacAddress = "5A:5B:5C:5D:5E:5F".parse().unwrap();
-
-    env.api
-        .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-NO-IP-001".into(),
-            bmc_ip_address: "".into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // No interface should exist for this MAC.
-    let mut txn = env.pool.begin().await?;
-    let interfaces = db::machine_interface::find_by_mac_address(&mut *txn, bmc_mac).await?;
-    assert!(
-        interfaces.is_empty(),
-        "should not create interface without bmc_ip_address"
-    );
-
-    Ok(())
-}
-
-/// Adding an expected power shelf with an external bmc_ip_address (not
-/// in any managed prefix) should create the interface on the
-/// static-assignments anchor segment.
-#[crate::sqlx_test()]
-async fn test_add_with_external_bmc_ip_uses_static_assignments(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let bmc_mac: MacAddress = "6A:6B:6C:6D:6E:84".parse().unwrap();
-    let external_ip = "10.50.1.150";
-
-    env.api
-        .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-EXT-001".into(),
-            bmc_ip_address: external_ip.into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // Add doesn't preallocate inline; mimic what site-explorer does on the next iteration --
-    // materialize the static BMC interface for this entity.
-    carbide_site_explorer::try_preallocate_one(
-        &env.pool,
-        bmc_mac,
-        external_ip.parse().unwrap(),
-        model::machine_interface::InterfaceType::Bmc,
-        "expected_power_shelf BMC",
-        None,
-    )
-    .await;
-
-    // Verify interface was created on the static-assignments segment
-    let mut txn = env.pool.begin().await?;
-    let interfaces = db::machine_interface::find_by_mac_address(&mut *txn, bmc_mac).await?;
-    assert_eq!(interfaces.len(), 1);
-
-    let iface = &interfaces[0];
-    assert!(iface.addresses.contains(&external_ip.parse().unwrap()));
-
-    let static_seg = db::network_segment::static_assignments(&mut txn).await?;
-    assert_eq!(
-        iface.segment_id, static_seg.id,
-        "external IP should be on the static-assignments segment"
-    );
-
-    Ok(())
-}
-
-/// Updating with bmc_ip_address that matches the existing address is a
-/// no-op -- the update succeeds without modifying the interface.
-#[crate::sqlx_test()]
-async fn test_update_with_matching_bmc_ip_is_noop(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let bmc_mac: MacAddress = "6A:6B:6C:6D:6E:81".parse().unwrap();
-    let bmc_ip = "192.0.1.191";
-
-    // Add expected power shelf with bmc_ip_address.
-    env.api
-        .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-NOOP-001".into(),
-            bmc_ip_address: bmc_ip.into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // Update with the same bmc_ip_address -- should succeed (no-op).
-    env.api
-        .update_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN-UPDATED".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-NOOP-001".into(),
-            bmc_ip_address: bmc_ip.into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    Ok(())
-}
-
-/// Updating with a different bmc_ip_address succeeds (updates expected
-/// data) but does not touch the interface if it already has addresses.
-/// Expected data is decoupled from managed state -- the interface IP
-/// can only be changed via assign-address / remove-address.
-#[crate::sqlx_test()]
-async fn test_update_with_different_bmc_ip_leaves_interface_alone(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let bmc_mac: MacAddress = "6A:6B:6C:6D:6E:82".parse().unwrap();
-    let original_ip = "192.0.1.192";
-
-    // Add expected power shelf with bmc_ip_address, then run the sweep so the static
-    // machine_interface row exists (the sweep is what materializes it).
-    env.api
-        .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-LEAVE-001".into(),
-            bmc_ip_address: original_ip.into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-    carbide_site_explorer::try_preallocate_one(
-        &env.pool,
-        bmc_mac,
-        original_ip.parse().unwrap(),
-        model::machine_interface::InterfaceType::Bmc,
-        "expected_power_shelf BMC",
-        None,
-    )
-    .await;
-
-    // Update with a DIFFERENT bmc_ip_address -- should succeed but
-    // not touch the interface (it already has an address).
-    env.api
-        .update_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-LEAVE-001".into(),
-            bmc_ip_address: "192.0.1.193".into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // Verify the interface still has the ORIGINAL IP.
-    let mut txn = env.pool.begin().await?;
-    let interfaces = db::machine_interface::find_by_mac_address(&mut *txn, bmc_mac).await?;
-    assert_eq!(interfaces.len(), 1);
-    assert!(
-        interfaces[0]
-            .addresses
-            .contains(&original_ip.parse().unwrap()),
-        "interface should still have the original IP, not the updated expected data IP"
-    );
-
-    Ok(())
-}
-
-/// Updating with bmc_ip_address should succeed if the interface exists
-/// but has no addresses (e.g., the address was expired/removed).
-#[crate::sqlx_test()]
-async fn test_update_with_bmc_ip_assigns_to_empty_interface(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let bmc_mac: MacAddress = "6A:6B:6C:6D:6E:83".parse().unwrap();
-    let relay: std::net::IpAddr = common::api_fixtures::FIXTURE_DHCP_RELAY_ADDRESS
-        .parse()
-        .unwrap();
-
-    // Create interface via DHCP, then remove its address.
-    let mut txn = env.pool.begin().await?;
-    let iface = db::machine_interface::validate_existing_mac_and_create(
-        &mut txn,
-        bmc_mac,
-        std::slice::from_ref(&relay),
-        None,
-        None,
-    )
-    .await?;
-    db::machine_interface_address::delete(&mut txn, &iface.id).await?;
-    txn.commit().await?;
-
-    // Add expected power shelf WITHOUT bmc_ip_address.
-    env.api
-        .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-EMPTY-001".into(),
-            bmc_ip_address: "".into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // Update with bmc_ip_address -- should succeed since interface has no addresses.
-    env.api
-        .update_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-EMPTY-001".into(),
-            bmc_ip_address: "192.0.1.194".into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // Verify the interface now has the static IP.
-    let mut txn = env.pool.begin().await?;
-    let addrs = db::machine_interface_address::find_for_interface(&mut txn, iface.id).await?;
-    assert_eq!(addrs.len(), 1);
-    assert_eq!(
-        addrs[0].address,
-        "192.0.1.194".parse::<std::net::IpAddr>().unwrap()
-    );
-    assert_eq!(
-        addrs[0].allocation_type,
-        model::allocation_type::AllocationType::Static
-    );
-
-    Ok(())
-}
-
-/// Updating with bmc_ip_address when no interface exists yet (device
-/// hasn't DHCP'd) should create a new interface with the static IP.
-#[crate::sqlx_test()]
-async fn test_update_with_bmc_ip_creates_interface_if_none_exists(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let bmc_mac: MacAddress = "6A:6B:6C:6D:6E:85".parse().unwrap();
-    let bmc_ip = "192.0.1.195";
-
-    // Add expected power shelf without bmc_ip_address.
-    env.api
-        .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-CREATE-001".into(),
-            bmc_ip_address: "".into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // No interface should exist yet.
-    let mut txn = env.pool.begin().await?;
-    let interfaces = db::machine_interface::find_by_mac_address(&mut *txn, bmc_mac).await?;
-    assert!(interfaces.is_empty());
-    txn.commit().await?;
-
-    // Update with bmc_ip_address -- should create a new interface.
-    env.api
-        .update_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-CREATE-001".into(),
-            bmc_ip_address: bmc_ip.into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // Verify interface was created with the static IP.
-    let mut txn = env.pool.begin().await?;
-    let interfaces = db::machine_interface::find_by_mac_address(&mut *txn, bmc_mac).await?;
-    assert_eq!(interfaces.len(), 1);
-    assert!(interfaces[0].addresses.contains(&bmc_ip.parse().unwrap()));
-
-    Ok(())
-}
-
-/// Updating without bmc_ip_address should not touch any machine
-/// interface -- only the expected device record is updated.
-#[crate::sqlx_test()]
-async fn test_update_without_bmc_ip_does_not_touch_interface(
-    pool: sqlx::PgPool,
-) -> Result<(), Box<dyn std::error::Error>> {
-    let env = create_test_env(pool).await;
-    let bmc_mac: MacAddress = "6A:6B:6C:6D:6E:86".parse().unwrap();
-    let bmc_ip = "192.0.1.196";
-
-    // Add with bmc_ip_address -- creates an interface.
-    env.api
-        .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "ADMIN".into(),
-            bmc_password: "PASS".into(),
-            shelf_serial_number: "PS-NOTOUCH-001".into(),
-            bmc_ip_address: bmc_ip.into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-    carbide_site_explorer::try_preallocate_one(
-        &env.pool,
-        bmc_mac,
-        bmc_ip.parse().unwrap(),
-        model::machine_interface::InterfaceType::Bmc,
-        "expected_power_shelf BMC",
-        None,
-    )
-    .await;
-
-    // Update without bmc_ip_address (just changing credentials).
-    env.api
-        .update_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
-            expected_power_shelf_id: None,
-            bmc_mac_address: bmc_mac.to_string(),
-            bmc_username: "NEW-ADMIN".into(),
-            bmc_password: "NEW-PASS".into(),
-            shelf_serial_number: "PS-NOTOUCH-001".into(),
-            bmc_ip_address: "".into(),
-            metadata: Some(rpc::forge::Metadata::default()),
-            rack_id: None,
-            bmc_retain_credentials: None,
-        }))
-        .await?;
-
-    // Verify the interface still has the original static IP.
-    let mut txn = env.pool.begin().await?;
-    let interfaces = db::machine_interface::find_by_mac_address(&mut *txn, bmc_mac).await?;
-    assert_eq!(interfaces.len(), 1);
-    assert!(
-        interfaces[0].addresses.contains(&bmc_ip.parse().unwrap()),
-        "interface should still have the original IP after update without bmc_ip_address"
-    );
-
-    Ok(())
-}
-
 /// When `bmc_retain_credentials` is set to true, the value should persist through
 /// add -> get round-trip via the RPC API.
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_add_expected_power_shelf_with_bmc_retain_credentials(
     pool: sqlx::PgPool,
 ) -> Result<(), Box<dyn std::error::Error>> {
     let env = create_test_env(pool).await;
     let bmc_mac: MacAddress = "7A:7B:7C:7D:7E:80".parse().unwrap();
 
-    env.api
+    env.api()
         .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
             expected_power_shelf_id: None,
             bmc_mac_address: bmc_mac.to_string(),
@@ -1298,7 +829,7 @@ async fn test_add_expected_power_shelf_with_bmc_retain_credentials(
         .await?;
 
     let retrieved = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelfRequest {
             bmc_mac_address: bmc_mac.to_string(),
             expected_power_shelf_id: None,
@@ -1317,7 +848,7 @@ async fn test_add_expected_power_shelf_with_bmc_retain_credentials(
 
 /// Verify that updating an expected power shelf without specifying `bmc_retain_credentials`
 /// preserves the existing value (and that COALESCE works).
-#[crate::sqlx_test()]
+#[sqlx_test]
 async fn test_update_expected_power_shelf_preserves_bmc_retain_credentials(
     pool: sqlx::PgPool,
 ) -> Result<(), Box<dyn std::error::Error>> {
@@ -1325,7 +856,7 @@ async fn test_update_expected_power_shelf_preserves_bmc_retain_credentials(
     let bmc_mac: MacAddress = "7A:7B:7C:7D:7E:81".parse().unwrap();
 
     // Create with bmc_retain_credentials = true.
-    env.api
+    env.api()
         .add_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
             expected_power_shelf_id: None,
             bmc_mac_address: bmc_mac.to_string(),
@@ -1340,7 +871,7 @@ async fn test_update_expected_power_shelf_preserves_bmc_retain_credentials(
         .await?;
 
     // Update without setting bmc_retain_credentials (None).
-    env.api
+    env.api()
         .update_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelf {
             expected_power_shelf_id: None,
             bmc_mac_address: bmc_mac.to_string(),
@@ -1355,7 +886,7 @@ async fn test_update_expected_power_shelf_preserves_bmc_retain_credentials(
         .await?;
 
     let retrieved = env
-        .api
+        .api()
         .get_expected_power_shelf(tonic::Request::new(rpc::forge::ExpectedPowerShelfRequest {
             bmc_mac_address: bmc_mac.to_string(),
             expected_power_shelf_id: None,
