+++ crates/api-core/tests/integration/main.rs
@@ -20,6 +20,9 @@ mod connected_device;
 mod credential_rotation;
 mod dhcp_lease_expiration;
 mod dpu_machine_inventory;
+mod expected_power_shelf;
+mod expected_power_shelf_crud;
+mod expected_power_shelf_static_address;
 mod expected_rack;
 mod explored_managed_host_find;
 mod explored_mlx_devices;
